feat(reorganization): Relocated components
