+++ client/src/components/general/InfoCard.tsx
