+++ client/src/pages/members/DepartmentOverview.tsx
@@ -7,8 +7,8 @@ import { makeStyles, createStyles } from "@mui/styles";
 import PageBar from "../../components/navigation/PageBar";
 import api from "../../utils/api";
 import { NavLink } from "react-router-dom";
-import InfoCard from "../../components/InfoCard";
-import DepartmentDialog from "./components/DepartmentDialog";
+import InfoCard from "../../components/general/InfoCard";
+import DepartmentDialog from "../../components/members/DepartmentDialog";
 import { DepartmentDetails, DepartmentMember, Director } from "../../types/membersTypes";
 import { showErrorMessage } from "../../utils/toastUtils";
 import { AuthContext } from "../../context/auth-context/AuthContext";
